Update .gitignore; restructure CE Package
