+++ .gitignore
@@ -1,2 +1,4 @@
 *.log
 crash.txt
+.idea
+out
