+++ game.log
@@ -260,3 +260,9 @@ Okt. 30, 2020 2:03:33 PM de.gurkenlabs.litiengine.environment.tilemap.xml.TmxMap
 WARNING: Tiled version 1.4.2 of map "Gurknukem" is greater than the supported version 1.2.x. Some features may not work.
 Okt. 30, 2020 2:03:33 PM de.gurkenlabs.litiengine.environment.tilemap.xml.TmxMap checkVersion
 WARNING: Tiled version 1.4.2 of map "Level1" is greater than the supported version 1.2.x. Some features may not work.
+Okt. 30, 2020 3:44:34 PM de.gurkenlabs.litiengine.environment.tilemap.xml.TmxMap checkVersion
+WARNING: Tiled version 1.4.2 of map "Arkham" is greater than the supported version 1.2.x. Some features may not work.
+Okt. 30, 2020 3:44:35 PM de.gurkenlabs.litiengine.environment.tilemap.xml.TmxMap checkVersion
+WARNING: Tiled version 1.4.2 of map "Gurknukem" is greater than the supported version 1.2.x. Some features may not work.
+Okt. 30, 2020 3:44:35 PM de.gurkenlabs.litiengine.environment.tilemap.xml.TmxMap checkVersion
+WARNING: Tiled version 1.4.2 of map "Level1" is greater than the supported version 1.2.x. Some features may not work.
