+++ src/calculationEngine/entitys/Attacks.java
@@ -1,7 +1,4 @@
-package calculationEngine.characters.beasts;
-
-import java.util.ArrayList;
-import java.util.List;
+package calculationEngine.entitys;
 
 public enum Attacks {
